+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 * `Versions`:
   * `Kotlin`: `1.8.20`
   * `Serialization`: `1.5.1`
-  * `MicroUtils`: `0.19.2`
+  * `MicroUtils`: `0.19.4`
   * `UUID`: `0.7.1`
 * `MVVM`:
   * Add `NavigationMVVMSingleActivity` as abstract class for single-activity oriented way of navigation